trigger on mods version check
This is to prevent a use case where a mod had new files created during
the user's playthrough, and the user updates that mod before deploying
or purging which would result in Vortex not being aware of the newly
created files.

This would only become observable to the user the next time he would
purge/remove the updated mod, leaving orphaned files behind.